consolidate interface test fixes
