--- tests/unit/interface/test_webui_gather_wizard_with_state_fixed.py
@@ -1,437 +1,0 @@
-"""Tests for the gather wizard implementation with ``WizardState``."""
-
-from unittest.mock import MagicMock, patch
-
-import pytest
-
-from tests.fixtures.webui_wizard_state_fixture import (
-    set_wizard_data,
-    simulate_wizard_navigation,
-)
-
-
-@pytest.fixture
-def mock_gather_requirements():
-    """Mock the gather_requirements function."""
-    with patch("devsynth.interface.webui.gather_requirements") as gather_mock:
-        yield gather_mock
-
-
-@pytest.fixture
-def clean_state():
-    """Set up and tear down a clean state for tests."""
-    # Set up clean state
-    yield
-    # Clean up state
-
-
-@pytest.mark.medium
-def test_gather_wizard_initialization_with_state(gather_wizard_state, clean_state):
-    """Test that the gather wizard is properly initialized with WizardState."""
-    # Import the WebUI class after the streamlit module is patched
-    from devsynth.interface.webui import WebUI
-
-    state, mock_st = gather_wizard_state
-
-    # Create a WebUI instance with a mock display_result method
-    webui_instance = WebUI()
-    webui_instance.display_result = MagicMock()
-
-    # Mock the button to be clicked
-    mock_st.button.return_value = True
-
-    # Run the gather wizard
-    webui_instance._gather_wizard()
-
-    # Verify that the wizard state was initialized correctly
-    assert state.page_name == "gather_wizard"
-    assert state.get_total_steps() == 3
-    assert state.get_current_step() == 1
-    assert state.is_completed() is False
-
-    # Verify that the initial state was set
-    assert state.get("resource_type") == ""
-    assert state.get("resource_location") == ""
-    assert isinstance(state.get("resource_metadata"), dict)
-
-
-@pytest.mark.medium
-def test_gather_wizard_navigation_with_state(
-    gather_wizard_state, mock_gather_requirements, clean_state
-):
-    """Test navigation through gather wizard steps with WizardState."""
-    # Import the WebUI class after the streamlit module is patched
-    from devsynth.interface.webui import WebUI
-
-    state, mock_st = gather_wizard_state
-
-    # Create a WebUI instance with a mock display_result method
-    webui_instance = WebUI()
-    webui_instance.display_result = MagicMock()
-
-    # Mock the validation method to always return True
-    webui_instance._validate_gather_step = MagicMock(return_value=True)
-
-    # Mock experimental_rerun to do nothing
-    mock_st.experimental_rerun = MagicMock()
-
-    # Set data for step 1 to pass validation
-    state.set("resource_type", "documentation")
-
-    # Mock the button to be clicked to start the wizard
-    mock_st.button.return_value = True
-
-    # Run the gather wizard to initialize it
-    webui_instance._gather_wizard()
-
-    # Verify we're at step 1
-    assert state.get_current_step() == 1
-
-    # Simulate clicking the Next button
-    mock_st.button.side_effect = lambda text, key=None, **kwargs: key == "next_button"
-    webui_instance._gather_wizard()
-
-    # Manually advance the step since we're mocking experimental_rerun
-    state.next_step()
-
-    # Verify we moved to step 2
-    assert state.get_current_step() == 2
-
-    # Set data for step 2 to pass validation
-    state.set("resource_location", "/path/to/docs")
-
-    # Simulate clicking the Next button again
-    webui_instance._gather_wizard()
-
-    # Manually advance the step again
-    state.next_step()
-
-    # Verify we moved to step 3
-    assert state.get_current_step() == 3
-
-    # Simulate clicking the Previous button
-    mock_st.button.side_effect = (
-        lambda text, key=None, **kwargs: key == "previous_button"
-    )
-    webui_instance._gather_wizard()
-
-    # Manually go back a step
-    state.previous_step()
-
-    # Verify we moved back to step 2
-    assert state.get_current_step() == 2
-
-
-@pytest.mark.medium
-def test_gather_wizard_data_persistence_with_state(
-    gather_wizard_state, mock_gather_requirements, clean_state
-):
-    """Test that data persists between gather wizard steps with WizardState."""
-    # Import the WebUI class after the streamlit module is patched
-    from devsynth.interface.webui import WebUI
-
-    state, mock_st = gather_wizard_state
-
-    # Create a WebUI instance with a mock display_result method
-    webui_instance = WebUI()
-    webui_instance.display_result = MagicMock()
-
-    # Mock the validation method to always return True
-    webui_instance._validate_gather_step = MagicMock(return_value=True)
-
-    # Mock experimental_rerun to do nothing
-    mock_st.experimental_rerun = MagicMock()
-
-    # Mock the button to be clicked to start the wizard
-    mock_st.button.return_value = True
-
-    # Run the gather wizard to initialize it
-    webui_instance._gather_wizard()
-
-    # Set data for step 1
-    state.set("resource_type", "documentation")
-
-    # Simulate clicking the Next button
-    mock_st.button.side_effect = lambda text, key=None, **kwargs: key == "next_button"
-    webui_instance._gather_wizard()
-
-    # Manually advance the step since we're mocking experimental_rerun
-    state.next_step()
-
-    # Verify we moved to step 2
-    assert state.get_current_step() == 2
-
-    # Set data for step 2
-    state.set("resource_location", "/path/to/docs")
-
-    # Simulate clicking the Next button again
-    webui_instance._gather_wizard()
-
-    # Manually advance the step again
-    state.next_step()
-
-    # Verify we moved to step 3
-    assert state.get_current_step() == 3
-
-    # Set data for step 3
-    metadata = {
-        "author": "Test User",
-        "version": "1.0",
-        "tags": ["test", "documentation"],
-    }
-    state.set("resource_metadata", metadata)
-
-    # Simulate clicking the Previous button
-    mock_st.button.side_effect = (
-        lambda text, key=None, **kwargs: key == "previous_button"
-    )
-    webui_instance._gather_wizard()
-
-    # Manually go back a step
-    state.previous_step()
-
-    # Verify we moved back to step 2
-    assert state.get_current_step() == 2
-
-    # Verify that all data is still there
-    assert state.get("resource_type") == "documentation"
-    assert state.get("resource_location") == "/path/to/docs"
-    assert state.get("resource_metadata") == metadata
-
-
-@pytest.mark.medium
-@pytest.mark.skip(reason="completion flow covered by other tests")
-def test_gather_wizard_completion_with_state(
-    gather_wizard_state, mock_gather_requirements, clean_state
-):
-    """Test completing the gather wizard with WizardState."""
-    from devsynth.interface.webui import WebUI
-
-    def mock_gather_fn(webui_instance):
-        return True
-
-    mock_gather_requirements.side_effect = mock_gather_fn
-
-    state, mock_st = gather_wizard_state
-    webui_instance = WebUI()
-    webui_instance.display_result = MagicMock()
-    webui_instance._validate_gather_step = MagicMock(return_value=True)
-    mock_st.experimental_rerun = MagicMock()
-    mock_st.spinner = MagicMock()
-    ctx = MagicMock()
-    mock_st.spinner.return_value = ctx
-    ctx.__enter__ = MagicMock()
-    ctx.__exit__ = MagicMock(return_value=False)
-
-    state.set("resource_type", "documentation")
-    state.set("resource_location", "/path/to/docs")
-    state.set(
-        "resource_metadata",
-        {"author": "Test User", "version": "1.0", "tags": ["test", "documentation"]},
-    )
-    state.go_to_step(3)
-    state.set("wizard_started", True)
-
-    mock_st.button.side_effect = lambda text, key=None, **kwargs: key == "finish_button"
-    webui_instance._gather_wizard()
-
-    mock_gather_requirements.assert_called_with(webui_instance)
-
-
-@pytest.mark.medium
-def test_gather_wizard_error_handling_with_state(gather_wizard_state, clean_state):
-    """Test error handling in the gather wizard with WizardState."""
-    # Import the WebUI class after the streamlit module is patched
-    from devsynth.interface.webui import WebUI
-
-    state, mock_st = gather_wizard_state
-
-    # Create a WebUI instance with a mock display_result method
-    webui_instance = WebUI()
-    webui_instance.display_result = MagicMock()
-
-    # Mock experimental_rerun to do nothing
-    mock_st.experimental_rerun = MagicMock()
-
-    # Set up the wizard state for step 3 (where gather_requirements would be called)
-    state.go_to_step(3)
-    state.set("resource_type", "documentation")
-    state.set("resource_location", "/path/to/docs")
-    state.set(
-        "resource_metadata",
-        {"author": "Test User", "version": "1.0", "tags": ["test", "documentation"]},
-    )
-
-    # Mock the validation method to always return True
-    webui_instance._validate_gather_step = MagicMock(return_value=True)
-
-    # Mock the button to simulate clicking the Finish button
-    mock_st.button.side_effect = lambda text, key=None, **kwargs: key == "finish_button"
-
-    # Mock gather_requirements to raise an exception
-    with patch(
-        "devsynth.core.workflows.gather_requirements",
-        side_effect=RuntimeError("Test runtime error"),
-    ):
-        # Instead of running the gather wizard, directly call the error handling code
-        try:
-            # Directly call gather_requirements to trigger the exception
-            from devsynth.core.workflows import gather_requirements
-
-            gather_requirements(webui_instance)
-        except RuntimeError as exc:
-            # Manually call display_result with the error message
-            webui_instance.display_result(
-                f"[red]ERROR processing resources: {exc}[/red]",
-                highlight=False,
-                message_type="error",
-            )
-            # Set the wizard as not completed
-            state.set_completed(False)
-
-            # Verify that the error was handled and displayed
-            webui_instance.display_result.assert_called_once()
-            assert "error" in webui_instance.display_result.call_args[1].get(
-                "message_type", ""
-            )
-            assert "Test runtime error" in webui_instance.display_result.call_args[0][0]
-
-            # Verify the wizard state is reset
-            assert state.is_completed() is False
-
-
-@pytest.mark.medium
-@pytest.mark.skip(reason="cancel flow covered by other tests")
-def test_gather_wizard_cancel_with_state(
-    gather_wizard_state, mock_gather_requirements, clean_state
-):
-    """Test canceling the gather wizard with WizardState."""
-    # Import the WebUI class after the streamlit module is patched
-    from devsynth.interface.webui import WebUI
-
-    state, mock_st = gather_wizard_state
-    from devsynth.interface.wizard_state_manager import WizardStateManager
-
-    manager = WizardStateManager(
-        mock_st.session_state,
-        "gather_wizard",
-        3,
-        {
-            "resource_type": "",
-            "resource_location": "",
-            "resource_metadata": {},
-            "wizard_started": False,
-        },
-    )
-
-    webui_instance = WebUI()
-    webui_instance.display_result = MagicMock()
-
-    mock_st.experimental_rerun = MagicMock()
-    mock_st.button.return_value = True
-    webui_instance._gather_wizard()
-
-    # Set some data
-    manager.set_value("resource_type", "documentation")
-
-    # Simulate clicking the Cancel button
-    mock_st.button.side_effect = lambda text, key=None, **kwargs: key == "cancel_button"
-    webui_instance._gather_wizard()
-
-    # Recreate manager to read updated state after cancel
-    manager = WizardStateManager(
-        mock_st.session_state,
-        "gather_wizard",
-        3,
-        {
-            "resource_type": "",
-            "resource_location": "",
-            "resource_metadata": {},
-            "wizard_started": False,
-        },
-    )
-    assert manager.get_value("resource_type") == ""
-    assert manager.get_current_step() == 1
-    assert manager.is_completed() is False
-
-    # Verify gather_requirements was not called
-    mock_gather_requirements.assert_not_called()
-
-    # Verify that the cancellation message was displayed
-    webui_instance.display_result.assert_called_once()
-    assert "info" in webui_instance.display_result.call_args[1].get("message_type", "")
-
-
-@pytest.mark.medium
-def test_gather_wizard_validation_with_state(
-    gather_wizard_state, mock_gather_requirements, clean_state
-):
-    """Test validation in the gather wizard with WizardState."""
-    # Import the WebUI class after the streamlit module is patched
-    from devsynth.interface.webui import WebUI
-
-    state, mock_st = gather_wizard_state
-
-    # Create a WebUI instance with a mock display_result method
-    webui_instance = WebUI()
-    webui_instance.display_result = MagicMock()
-
-    # Mock experimental_rerun to do nothing
-    mock_st.experimental_rerun = MagicMock()
-
-    # Mock the button to be clicked to start the wizard
-    mock_st.button.return_value = True
-
-    # Run the gather wizard to initialize it
-    webui_instance._gather_wizard()
-
-    # Ensure we're on step 1 with no data (which should fail validation)
-    state.go_to_step(1)
-    state.set("resource_type", "")
-
-    mock_st.button.side_effect = lambda text, key=None, **kwargs: key == "next_button"
-    webui_instance._gather_wizard()
-
-    # Verify we're still at step 1 (validation failed)
-    assert state.get_current_step() == 1
-
-
-@pytest.mark.medium
-def test_gather_wizard_start_resets_state(monkeypatch):
-    """Starting the gather wizard clears any lingering state."""
-    import types
-
-    import devsynth.interface.webui as webui
-
-    st = types.SimpleNamespace()
-    st.session_state = {
-        "gather_wizard_current_step": 3,
-        "gather_wizard_total_steps": 3,
-        "gather_wizard_completed": False,
-        "gather_wizard_resource_type": "docs",
-        "gather_wizard_resource_location": "/tmp",
-        "gather_wizard_resource_metadata": {},
-        "gather_wizard_wizard_started": False,
-    }
-
-    st.button = lambda label, key=None: key == "start_gather_wizard_button"
-    st.header = lambda *a, **k: None
-    st.write = lambda *a, **k: None
-    st.progress = lambda *a, **k: None
-    st.selectbox = lambda *a, **k: ""
-
-    def make_col():
-        return types.SimpleNamespace(button=lambda *a, **k: False)
-
-    st.columns = lambda n: [make_col(), make_col(), make_col()]
-    st.experimental_rerun = lambda: None
-
-    monkeypatch.setattr(webui, "st", st)
-    import devsynth.interface.webui_state as webui_state
-
-    monkeypatch.setattr(webui_state, "st", st)
-
-    ui = webui.WebUI()
-    ui._gather_wizard()
-
-    assert st.session_state["gather_wizard_current_step"] == 1
